feat: Add manualSplit flag to trips
Introduces a `manualSplit` boolean field to the `Trip` type and database schema.
This flag indicates if a trip was manually split by the user into separate departure and return legs.
When a trip is manually split, the `manualSplit` flag is set to `true` on the newly created return trip.
Additionally, when two manually separated trips (departure and return) are manually grouped back together by the user, the `manualSplit` flag on the resulting combined trip is reset to `false`.
This flag will be used to prevent the automatic grouping logic from merging trips that the user has intentionally separated.